// frontend/src/api/model/diagnosisLevel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type DiagnosisLevel = typeof DiagnosisLevel[keyof typeof DiagnosisLevel];


export const DiagnosisLevel = {
  CHAPTER: 'CHAPTER',
  BLOCK: 'BLOCK',
  CATEGORY: 'CATEGORY',
  SUBCATEGORY: 'SUBCATEGORY',
} as const;