// frontend/src/api/model/diagnosisStage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Hospital Internment API
 * OpenAPI spec version: 0.1.0
 */

export type DiagnosisStage = typeof DiagnosisStage[keyof typeof DiagnosisStage];


export const DiagnosisStage = {
  ADMISSION: 'ADMISSION',
  DISCHARGE: 'DISCHARGE',
} as const;